+++ lib/src/spotify_base.dart
@@ -30,6 +30,8 @@ abstract class SpotifyApiBase {
   Categories get categories => _categories;
   Me _me;
   Me get me => _me;
+  Shows _shows;
+  Shows get shows => _shows;
   oauth2.Client get client => _client;
 
   SpotifyApiBase.fromClient(FutureOr<http.BaseClient> client) {
@@ -45,6 +47,7 @@ abstract class SpotifyApiBase {
     _search = Search(this);
     _audioFeatures = AudioFeatures(this);
     _categories = Categories(this);
+    _shows = Shows(this);
   }
 
   SpotifyApiBase(SpotifyApiCredentials credentials, [http.BaseClient httpClient]) : this.fromClient(_getOauth2Client(credentials, httpClient));